# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FACEIT Demo Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type DemoStatus = typeof DemoStatus[keyof typeof DemoStatus];


export const DemoStatus = {
  pending: 'pending',
  ready: 'ready',
  error: 'error',
} as const;

export interface Demo {
  id: number;
  filename: string;
  /** @nullable */
  matchId?: string | null;
  map: string;
  importedAt: string;
  status: DemoStatus;
  /** @nullable */
  team1Name?: string | null;
  /** @nullable */
  team2Name?: string | null;
  playerCount?: number;
}

export type DemoDetailStatus = typeof DemoDetailStatus[keyof typeof DemoDetailStatus];


export const DemoDetailStatus = {
  pending: 'pending',
  ready: 'ready',
  error: 'error',
} as const;

export interface Player {
  id: number;
  steamId: string;
  name: string;
  slot: number;
  team: number;
}

export interface DemoDetail {
  id: number;
  filename: string;
  /** @nullable */
  matchId?: string | null;
  map: string;
  importedAt: string;
  status: DemoDetailStatus;
  /** @nullable */
  team1Name?: string | null;
  /** @nullable */
  team2Name?: string | null;
  players: Player[];
}

export interface VoicePreset {
  label: string;
  description: string;
  command: string;
  indicesLow: number;
  indicesHigh: number;
}

export interface DemoImportInput {
  filePath: string;
  map?: string;
  team1Name?: string;
  team2Name?: string;
  /** ISO 8601 date from the file's last-modified time */
  importedAt?: string;
}

export interface MapCount {
  map: string;
  count: number;
}

export interface DemoStats {
  totalDemos: number;
  readyDemos: number;
  totalPlayers: number;
  maps: MapCount[];
}

export interface Settings {
  /** @nullable */
  cs2Path: string | null;
  watchedFolders: string[];
  autoImport?: boolean;
  replaysSubfolder?: string;
}

export interface SettingsInput {
  cs2Path?: string;
  watchedFolders?: string[];
  autoImport?: boolean;
  replaysSubfolder?: string;
}