# lib/api-zod/src/generated/types/demoImportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FACEIT Demo Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface DemoImportInput {
  filePath: string;
  map?: string;
  team1Name?: string;
  team2Name?: string;
  /** ISO 8601 date from the file's last-modified time */
  importedAt?: string;
}